# backend/documentation/openapi.yaml
openapi: 3.1.0
info:
  title: BG GOLD Attendance API
  version: 0.1.0
  description: Organization-scoped REST API. All timestamps are UTC RFC 3339 and attendance time is server-authoritative.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      security: []
      summary: Create an access and rotating refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200": { $ref: "#/components/responses/TokenPair" }
        "401": { $ref: "#/components/responses/Error" }
  /audit-logs:
    get:
      summary: List organization-scoped immutable audit activity
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: action, schema: { type: string } }
        - { in: query, name: resourceType, schema: { type: string } }
        - { in: query, name: actorUserId, schema: { type: string, format: uuid } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { in: query, name: cursor, schema: { type: string } }
      responses:
        "200": { description: Audit items and optional top-level nextCursor }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /auth/refresh:
    post:
      security: []
      summary: Rotate a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/TokenPair" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password/forgot:
    post:
      security: []
      summary: Request a non-enumerating password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        "202":
          description: Generic response regardless of account existence
        "422": { $ref: "#/components/responses/Error" }
  /auth/password/reset:
    post:
      security: []
      summary: Consume a single-use reset token and revoke existing sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        "200": { description: Password updated and sessions revoked }
        "422": { $ref: "#/components/responses/Error" }
...
  /me/active-organization:
    post:
      summary: Switch organization and rotate the active session token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId]
              properties:
                organizationId: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/TokenPair" }
        "403": { $ref: "#/components/responses/Error" }
  /me/shifts:
    get:
      summary: List published shifts assigned to the active membership
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        "200": { $ref: "#/components/responses/ShiftList" }
  /me/attendance-policy:
    get:
      summary: Resolve effective attendance policy
      parameters:
        - in: query
          name: sectionId
          schema: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Policy" }
        "409": { $ref: "#/components/responses/Error" }
...
  /me/attendance/history:
    get:
      summary: List append-only attendance events
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - {
            in: query,
            name: limit,
            schema: { type: integer, minimum: 1, maximum: 100 },
          }
        - { in: query, name: cursor, schema: { type: string }, description: Opaque cursor returned as nextCursor by the preceding page. }
      responses:
        "200": { description: Attendance event list with an optional top-level nextCursor }
        "400": { $ref: "#/components/responses/Error" }
  /me/attendance/events/{eventId}/evidence:
    get:
      summary: Read evidence for an attendance event owned by the active employee
      parameters:
        - in: path
          name: eventId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Own event, location, device, verification, and five-minute private photo URL }
        "404": { $ref: "#/components/responses/Error" }
...
  /attendance/actions:
    post:
      summary: Submit an idempotent attendance state transition
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, maxLength: 160 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceActionInput" }
      responses:
        "201": { description: New event recorded }
        "200": { description: Previously stored idempotent response }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /me/open-shifts:
    get:
      summary: List published open shifts available to the active employee
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        "200": { description: Open shifts with the employee request status }
  /shifts/{shiftId}/requests:
    post:
      summary: Request an available open shift
      parameters:
        - { in: path, name: shiftId, required: true, schema: { type: string, format: uuid } }
      responses:
        "201": { description: Pending shift request created }
        "409": { $ref: "#/components/responses/Error" }
  /shift-requests:
    get:
      summary: List organization shift requests for manager review
      responses:
        "200": { description: Shift request queue }
        "403": { $ref: "#/components/responses/Error" }
  /shift-requests/{requestId}/decision:
    post:
      summary: Approve or reject a shift request
      parameters:
        - { in: path, name: requestId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Decision stored; approval creates an assignment }
        "409": { $ref: "#/components/responses/Error" }
  /sections/{sectionId}/dynamic-qr:
    post:
      summary: Issue a signed single-use attendance QR for an active section
      parameters:
        - in: path
          name: sectionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: QR token valid for 45 seconds
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      { data: { $ref: "#/components/schemas/DynamicQR" } }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /attendance/requests:
    get:
      summary: List organization attendance requests for review
      parameters:
        - in: query
          name: status
          schema: { enum: [PENDING, APPROVED, REJECTED, WITHDRAWN] }
      responses:
        "200": { $ref: "#/components/responses/AttendanceRequestList" }
        "403": { $ref: "#/components/responses/Error" }
  /attendance/records:
    get:
      summary: List organization attendance records with effective decisions and latest corrections
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - {
            in: query,
            name: membershipId,
            schema: { type: string, format: uuid },
          }
        - {
            in: query,
            name: limit,
            schema: { type: integer, minimum: 1, maximum: 500 },
          }
      responses:
        "200": { description: Organization-scoped attendance record list }
        "403": { $ref: "#/components/responses/Error" }
  /attendance/events/{eventId}/evidence:
    get:
      summary: Read organization-scoped attendance evidence with a five-minute private photo URL
      parameters:
        - in: path
          name: eventId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Attendance event, location, device, integrity, face-verification, and private attachment metadata }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /attendance/timesheets:
    get:
      summary: Compute organization timesheets using effective corrections and break rounding
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - {
            in: query,
            name: membershipId,
            schema: { type: string, format: uuid },
          }
      responses:
        "200": { description: Daily gross, actual break, rounded break, and net work minutes }
        "403": { $ref: "#/components/responses/Error" }
  /reports/attendance.csv:
    get:
      summary: Export effective attendance records as UTF-8 CSV
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: membershipId, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Attendance CSV including correction markers }
        "403": { $ref: "#/components/responses/Error" }
  /reports/timesheets.csv:
    get:
      summary: Export daily gross, break, and net timesheets as UTF-8 CSV
      parameters:
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
        - { in: query, name: membershipId, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Timesheet CSV using policy-rounded breaks }
        "403": { $ref: "#/components/responses/Error" }
  /attendance/requests/{requestId}/decision:
    post:
      summary: Append an approval or rejection decision
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceDecisionInput" }
      responses:
        "200": { description: Decision stored and effective state updated }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /attendance/corrections:
    post:
      summary: Append a correction without mutating the original event
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceCorrectionInput" }
      responses:
        "201": { description: Correction event and link created }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /leave-types:
    get:
      summary: List active organization leave types
      responses:
        "200": { $ref: "#/components/responses/LeaveTypeList" }
    post:
      summary: Create an organization leave type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLeaveType" }
      responses:
        "201": { description: Leave type created }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /leave-balances:
    post:
      summary: Set an employee leave entitlement for a calendar year
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetLeaveBalance" }
      responses:
        "200": { description: Entitlement stored }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /me/leave-balances:
    get:
      summary: List the active employee's leave balances
      parameters:
        - { in: query, name: year, schema: { type: integer, minimum: 2000, maximum: 2200 } }
      responses:
        "200": { $ref: "#/components/responses/LeaveBalanceList" }
  /me/leave-requests:
    get:
      summary: List the active employee's leave requests
      responses:
        "200": { $ref: "#/components/responses/LeaveRequestList" }
    post:
      summary: Request leave and reserve the required balance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLeaveRequest" }
      responses:
        "201": { description: Pending leave request created }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /me/leave-requests/{requestId}/withdraw:
    post:
      summary: Withdraw a pending own leave request and release its balance
      parameters:
        - { in: path, name: requestId, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Request withdrawn }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /leave-requests:
    get:
      summary: List organization leave requests for manager review
      parameters:
        - { in: query, name: status, schema: { enum: [PENDING, APPROVED, REJECTED, WITHDRAWN] } }
      responses:
        "200": { $ref: "#/components/responses/LeaveRequestList" }
        "403": { $ref: "#/components/responses/Error" }
  /leave-requests/{requestId}/decision:
    post:
      summary: Approve or reject leave and update reserved balances atomically
      parameters:
        - { in: path, name: requestId, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveDecisionInput" }
      responses:
        "200": { description: Append-only decision stored }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /attachments/attendance-selfie:
    post:
      summary: Upload a private attendance selfie with 180-day retention
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Finalized private attachment }
        "400": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
  /attachments/face-image:
    post:
      summary: Upload a private face image with 30-day retention
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: Finalized private face image }
        "400": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
  /me/face/enroll:
    post:
      summary: Enroll the active membership with the configured face provider
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FaceImageInput" }
      responses:
        "200": { description: Active face enrollment }
        "503": { $ref: "#/components/responses/Error" }
...
  /claims:
    get:
      summary: List organization claims for manager review
      parameters:
        - { in: query, name: status, schema: { enum: [PENDING, APPROVED, REJECTED, WITHDRAWN] } }
      responses:
        "200": { description: Organization claim list }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Envelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string }
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        organizationId: { type: string, format: uuid }
...
  responses:
    Error:
      description: Stable machine-readable error
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
        }
    TokenPair:
      description: Token pair
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - type: object
                properties:
                  { data: { $ref: "#/components/schemas/TokenPairData" } }
    ShiftList:
      description: Shift list
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/Envelope" } },
        }
    Policy:
      description: Effective policy
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/Envelope" } },
        }
...